Added Combining Exponent Rules - Now Working on Blanking Out Values
I added the ability to combine the exponent rules for the game, and now I'm working on blanking out values. I need to make it so that the questions are solvable, which right now the program doesn't check for when removing values. An easy way of doing it is having the program only ever remove things from one side of the equal sign, which may be something I do, even only as a quick implementation. Combining exponent rules would obfuscate this limitation, but I don't know if I want to just settle with that (though I assume I can by just generating multiple terms of the same exponent rule type). Actually, I think I may just do it that way... With that said, I need to blank out the values before I combine the questions so that the program knows what to remove to keep the question solvable.