Added Progress Bar, Loading Screen + Filling in Values Works
I added in the progress bar and the loading screen, but I'm not using either of them yet. I also fixed the question-answer system since the question string wasn't actually getting overwritten like it should. Putting in the correct value works, but the display isn't updated, and nothing happens when you get the question correct.