Finished Blanking Out System
Finished the system for blanking out values to answer the question with. It's kind of hard to read, so I need to sort that out later. I'm going to need to set up the keyboard system soon just so that I can properly fill the questions. I might also rename some member variables to reduce confusion.