Keypad Puzzle + Index Fixes + Rule Adjustments
I added the keypad puzzle, which included the number and keypad images. It seems to work properly, though I need to figure out the resolutions (it looks blurry).

I fixed the indexes of the values so that the proper values are replaced when a correct value is put in. As for the rules, I made some adjustments to make sure the questions are always solvable, and that certain cases are ignored. However, when I tried setting all but the zero rule to not be used, I still got the base exponent rule, so I need to fix that. However, that fix will come later. I'll probably set up the points system and question refresh next just to complete the game loop for the puzzle controls.