MatchInfo Instantiation Works
Instantiating using MatchInfo works. It only generates the puzzle mechanics, with everything else already being set up. There's still more I have to add to it, but the basic match setup is working for now. Speaking of which, I do need to add buttons for going back to the world and settings and such, but that will come later.

For now, I think I'll work on the game world next for moving into the match scene.