Created DebugWorldScene, and Added More Files from Other Project
I added a debug world scene, but I haven't done anything with it yet. I also added the textbox, page, and language marker scripts from another project to be used for later. I also added the scenes to the build list. Make sure to remove the debug scenes from the build list in the end.