Added Challenger Event
Added an event to enable a challenger after certain challengers have been beaten. I also added parameters to check if the challenger is available. Maybe I should add a variable to see if the challenger should be hidden when unavailable. I know rematches aren't allowed, but maybe for the Itch.io version I'll do that.