Area Event Complete + Working on Tutorial
The area event is complete. It turns on the forward button when all the connected challengers are defeated. I have enough to set up the three areas for the game world, but I won't do that just yet.

I'm now working on setting up the tutorials for the world and match scenes. I've split them between two files for ease of access. I'll probably set up some default text to test it, and a UI object to enable/disable as needed.