More Work on Tutorial - Set Up Text Box
Set up a basic text box for the tutorial. I should have the tutorial be set to call the functions for when the tutorial starts and when the tutorial ends, but I don't know if I should do this in the tutorial parent script or the children scripts. Next I'll probably work more on the tutorial to make sure it works properly with the matches and the world.