Working on the Tutorial Set Up
Working on setting up the tutorial as a singleton. The GameUI script is used to save the textbox for the tutorial, and will also hold the settings interface. The GameplayManager script adds the callbacks to the tutorial text box so that it can be opened and closed while also calling the appropriate functions.