Power System and Skip Implemented
Created the power system for the game. The only power is a debug power that does nothing, but the game is set up to allow the use of powers. I'm getting close to being able to set up the world and match scenes for the main game.

The skip function also works. I still need to add a cooldown for using it, but the function at least works.