Implemented Bubble Mechanic
Implemented the bubble mechanic and made some related fixes to the puzzle system. I needed to scale down the bubbles, which may be something that I adjust in the final game by changing the size of the camera.

I'm going to work on the sliding puzzle next, and change the random value function so that it limits the values generated.