Working on the World to Match Scene System
Working on having the user go from the world to the match scene, which appears to work now. I haven't checked going back from the match scene though, so that needs to be worked on.

The mouse projection system for selecting items in the puzzle space is broken. I think I'll convert the hit point to a percentage of the object, then use that in the camera's viewport space to get the camera raycast position. Hopefully that'll work.