Implemented World to Match Jumping (And Vice-Versa)
The implemented the ability to go to the world to the match, and vice versa. I haven't tried it yet, but as it stands I think the game is playable from start to finish. I still need to set up menus, power selection, and power unlocking, but once all that's done I can work on some of the art assets. I might make some small tweaks first though since there are still some small things that I need to adjust (e.g., the equation display).