Working on the Power Selection Menu and Power Unlocking
Working on the power selection and unlocking. Unlocking appears to work, but there's no notification to show that the power has been unlocked. I'll probably just make a small prompt instead of making it a menu the user has to click through.

As for the power selection menu, I'm currently working on a script for it. I have no UI elements set up though.