Working on Power Menu + Saving and Loading
I was working on the power menu, then I realized I never planned out the layout. So I'm going to go do that before I continue.

In the meanwhile, I worked on the saving and loading system, but I haven't tested it yet. It also doesn't account for some parts of the game data, such as tutorials (which haven't been implemented yet). I should probably test this soon.

I also moved around and renamed some files, but that's not a big thing to detail.