Working on the Power Menu
I'm working on the power menu. Equipping the selected power and unequipping the current power work fine, but the power display is kind of iffy. I probably won't allow the player to open it if they have no powers anyway, but I want to fix that up. After that, I need to program the tutorial, test the saving, and work on the language file. I really want to finish all major programming requirements before the end of the month so that I have a good 2-3 weeks for the art and audio assets.