Finished Basic Power Menu
The basic power menu works. The visuals are rough and everything, but mechanically it's functional. I also added an object to the world scene that blocks the world colliders when a menu is up, but I probably need to add some safe guards to make sure this isn't accidently left on or something. But that's for later.

Once I get the tutorial code figured out I think I'll start making some art assets for the game. The other things I have to code are either smaller additions or things that I could implement quickly.