Tutorial Triggers Work in World and Match Scenes
The tutorial triggers work in the game world and in the game match. I still need to do some more testing to make sure that there's no glitches or conflicting things available when a tutorial is running, but that's functional.

I'm probably going to do some art assets for the game soon.