Updated Exponent Rules
I updated the exponent rule names so that they match the enums, which use the official titles. I also made the adjustments in the world scene for it, but the game still seems to include the regular exponent rule regardless of what the settings are. I need to fix that.