Build Check for Project
Successfully built the project for the game. I just did this to make sure the project could actually be built. There were some UnityEditor inclusions that needed to be removed for the game to run.

I set the code stripping to high, but made no other optimizations, such as removing the baseline LOL content. The game is currently 14.9 MB, which is well under the 30 MB limit, and around the 15 MB recommended project size. Not all the audio is in yet (which takes up a lot of space), so hopefully after all the optimizations you can still keep the project around 15 MB. You may change plans if you can't, but I don't think you would have to.