Fixed Rule Selection Bug + Working on Difficulty and Box Placement
I fixed the bug where the rule selection would give you rules that you didn't select.

I added in some difficulty settings for AI speed. I haven't tested this much, but it's something.

I'm also working on showing the empty box for the missing values, instead of showing "[ ]". I need to implement the prefab so that it can be placed in the scene.