Added in Boss Music, Changed Results, and Added Value Box Image
Added in the boss music, and replaced the results music with a different cut of the song. I also added the value box image prefab, but I haven't done anything with it yet. I may not even bother using it in the end (as it would be a pain to implement), but that awaits to be seen.