Updated Some Puzzle Sprites
Updated some of the puzzle sprites for the game. Speaking of which, it turns out the sprite has a built-in sorting option. I didn't need to attach the sorting group to these sprites, so I should go back and fix that.

All that's left is the character art, and the power icons. I should also make some kind of background to be used for the UI elements since I don't have one that fits the setting. I should also make a button sprite that fits the setting, but I may hold off on that.