Audio Bug Fix + Created the Rest of the Powers
I fixed a bug where the audio would be set lower than it's supposed to be when the settings menu is opened.

I created the rest of the powers for the game, and added them to the power info script. There is the block power that I thought about doing, but that would be more complex to program, which is why I didn't bother. Maybe I will in the future, I don't know.

The powers have not been tested as of yet, so I need to do that.