Fixed Puzzle and Power Bug + Added Power Unlock Animation
Fixed a bug where the puzzle wouldn't be running properly when you started another match, due to the game still being 'paused' internally. I also fixed a power bug where the power bar display for P2 wasn't working (the update was being sent to P1's bar by accident).

I added a power unlock animation for when the player gets a power. It isn't anything special, but it's something I can work on later. I also set the powers for the challengers in the world, though the AI is still not programmed to use them.