Added Points Block Power + Bubble Mechanic Adjustments
I added the points block power to the game, though I don't want to use it unless I can make some sort of animation for it to better indicate what's happening.

I adjusted the bubble mechanic to spawn more bubbles. I still need to adjust the value randomization so that the odds of the correct value being provided is more likely. I'll likely do that next since the bubble puzzle takes a while with the full random set of values.