Random Value Change + Fixed Powers + Fixed Area Transition
I've edited the random value generator so that it's more likely to give the user the value they need.

I fixed the powers so that they can be equipped and used properly. I also adjusted their fill and depletion rates, though I should probably make more changes there. The AI is still not programmed to use powers, and I haven't set up equation term counts yet.

The area transition has been adjusted so that it can be skipped. I should probably program it to prevent button pushes too.

I also changed some settings so that the patterns could properly be used as tiles, but that's a minor change.