Finished Tutorial Translations + Save System Works?
I finished implementing the tutorial translations, but the actual triggers are not set up yet.

I tested the save system for the game, and it appears to work. The feedback text isn't showing up for some reason, so I need to fix that, but saving and loading appears to work. I want to get the feedback text fixed before I try a full playthrough. I also think the points scaling should be adjusted, because matches are going too slow.