Tested WebGL Build
I tested the WebGL build in the harness, and it works. The save text showed up there, so it seems like there isn't any issues there. I guess the editor just saves that fast or something?

Anyway, I can't build for Windows due to the WebGL integration. It appears there might be a workaround, but the example scripts still prevent a Windows build from being made. I'll probably just leave it.

You should try to do a full playthrough of the game to see if it's possible.