Working on Game Playthrough
Testing out a gameplay playthrough. For some reason the defeated flag is not getting set properly on some challengers. Maybe it's because they aren't active at the time or something? I need to figure out what the problem is there.

Returning from a saved game seems to fix it, but when going through gameplay normally it doesn't work consistently.