Added SubmitProgress + General Game Test
I added in SubmitProgress calls since I realized I was missing those. I'm also still doing a general game playthrough test. For some reason, the enemies past the first 3 are not getting marked as defeated like they're supposed to. If you load a saved game, it seems to work, but not for a continuous game, and I don't know why.

Also, when facing the boss for some reason the boss doesn't have values to pick from, and it causes the game to error out. Maybe it's because of the pinball puzzle, or all the puzzle types being enabled, but I don't know. This only happens with specifically the boss, so some setting must be screwing up. Either way, I need to figure out why these are happening. I still need to put in the tutorial triggers, and I would like to complete my saved game before doing so.

Oh, and for some reason the game allowed you to progress to the next area even though the enemy wasn't marked as beaten. I assume that's tied into the enemy defeated parameter not being set properly.