Game Can Be Completed, but Defeated Parameter is Still Wonky
The game can be completed, and I've added a save file flag to check for the game being completed. I added the completed parameter late, so I haven't tested it yet, but it should work.

The defeated parameter still doesn't consistently work, and I don't know why. I think it has something to do with setting the area, or maybe the area event or something? I'm not sure. If anything, I still need to figure this out.