Made Small Changes to the Power Menu
Made some small changes to the power menu so that the equip button is disabled if the player has the selected power. It should be noted that this doesn't fully work properly (doesn't work when you first open the menu), so I need to fix that. I may leave that for the second build of the game though.

I also need to disable the world colliders when the player opens a menu. That bug really needs to be fixed.

I really need to add the tutorial text as well. That's the last major mechanic I'm missing.