Added Skip Delay and Reset Match Function
I added a skip delay to the game, though I need to make sure it doesn't interfere with any other functions like windows opening or tutorials being shown. It seems alright so far.

I added a reset match function for replaying the game. I don't know if I'll use it, since it's complex and I haven't tested it yet, but it'll probably be a complaint. Maybe I should leave that for build 2.