Match Canvas Changes + Rematch + Power Fixes
I made changes to the match canvas so that the player has more buttons, and some functions are streamlined (e.g., the Match End buttons refer to Match End, not the Match UI).

I programmed the rematch feature. It appears to work, but I haven't tested it with the other puzzle types. I also need to make sure to change the reticle position back to its default when I do it.

I fixed the points up and points down powers, since the multipliers were set to 1.0, meaning they did nothing. I haven't done a comparison yet.