Added Reticle Reset Pos + Fixed Match Audio + Added Char Counts
I added a reset position for the reticle. Now when the match is reset the reticle goes back to its original place. Make sure to disable the sprite in the final game.

I fixed the match audio so that the match BGM plays when the match restarts. I also made some adjustments to how it works.

I added some char counts to the language file. I still need to add more though. After all that, the game should be about good for the first build.