Finished Char Counts
I finished adding the char counts to the language file, and I made some adjustments for them.

Now I'm making sure all the buttons can support 15 characters. The text isn't meant to be in all caps, so the buttons will only support 15 chars max since the content shouldn't exceed the width of the button anyway. I'll try to have a better system for this when I make build 2.