Rebuilt the Game
I rebuilt the game. Right now it's 17.9 MB. I should be able to get it smaller by compressing some files.

I think the challenger settings may be getting screwed up by having LoadWorldManager() in the Start() function. I'll move it into the PostStart() function and see if that fixes it.