Full Playthrough Test, Some Adjustments
I did a full playthrough test. I was able to beat it, but the final boss was too difficult. I decided to increase the wait time, and reduce the number of equations the boss throws at you.

I tried moving the world load to another place, but it didn't fix anything. I think the game might be reloading the same save data, which could be causing problems. I'm going to check that and see if that's what's happening. If it is, then I've found the problem.