Fixed Major Game Save Glitch
I fixed a major bug that was causing the game not to progress properly. Basically, the game kept reloading the same save data every time it returned to the world scene, which was overriding the most recent, new data. I fixed the issue, so now the game progresses like normal. That was a big problem, and now it's fixed.

There are small things I still need to do before the deadline, but the game's at least in a workable state now.