Updated Power and Exponent Rule Information
I've updated the power and exponent rule information with rewrites. I'm going to do the tutorial next, but that'll be a bit more since I need to adjust the page counts.

I also changed the trigger for the final match tutorial since it should appear at the end of area 2, rather than at the start of area 3.