Fixed Match Menu Bug
Fixed a match menu bug where the match would get unpaused when match end is active if the settings were opened.

The game currently has an uncompressed build size of 15.3 MB when building for fastest build time. It's recommended that the game be around 15 MB (30 MB is the actual limit), and I don't think the game will get bigger, so this is a good size.

I'll do some more playthroughs tomorrow before turning it in.