Fixed AI Error + Changed Twist + Fixed Title Scene Errors
Fixed an AI error where it would check for missing values in the question when there are none.

I adjusted the cooldown speed of the twist power.

I fixed the backgrounds in the title scene so that they used the same patterns as everywhere else. I also fixed some display errors for the licenses window.