Adjusted UI, Fixed Minus Power, Tutorial Change, and Text Box Change
I adjusted the UI to resize a button I forgot to resize. I also made some small text changes for the tutorial.

I fixed the minus power so that it actually reduces the target's points instead of increasing them.

I changed the text box so that the back button is disabled on the first page.