Language File Update, Updated Challenger Icons, and Symbols
The language file has been updated with a new version from a build review.

I updated the challenger icons with new versions, but I decided to keep them at the same size due to some uses having them be over 128x128.

I replaced the power icons with versions that are 128x128 because no uses go above that size. I also added a skip button of the same size but I don't know if I'll use it.