Added in Shortened Music, TTS Fix
For space reasons, I shortened the songs for the game, and I have updated them accordingly. I'm going to build the game and see if it's below 30 MB uncompressed, as I had to remove the boss theme to accomplish this last time.

I also took a look at the TTS volume controls, and it appears to work fine. I don't know why it didn't work during the content reviews.