Added in New Prompt Images and Adjusted Difficulties
I added in the new prompt images, though I may replace them with silhouettes or something like that. I don't like the way the prompt images look.

I adjusted the difficulty settings for the game, making the AI easier, or at least that was my intention. I haven't done a full playthrough with these changes, so I'll need to come back to that.

The game's current size is 27.283014 MB. I still need to remove the unused text for the challenger UI, and update some of the art assets. Hopefully nothing I add will make it go over 30 MB uncompressed.