Updated Character Prompts, Set Custom Challenger UI Colours
I updated the challenger prompts to use the full renders. I also set custom colours for each character to use for the UI. The text is readable, so hopefully the custom colours shouldn't be a problem for the client.

I also added a higher resolution render of the tutorial character to use on the results screen.