Updated Backgrounds, Updated Pinball
I updated the title and match backgrounds. I didn't change much, but they've been adjusted a little.

I updated the pinball ball and gate. For the ball, I changed the sprite so that the top and bottoms are clearer. For the gate, I changed the max weight, and I made it so that the gate changes colour based on its current weight. I also removed the ball hit sound effect because it was annoying.