Adjusted Results Screen, Working on Text Scaler
I adjusted the results screen to include a win message and an image of the tutorial character. I am also working on the text scaler.

I wanted to use the button colour patterns, but they don't look good with the text. I might just stick with Unity's default buttons, which looks lazy, but visually better.