Added Power, Skip Diagrams and TTS for Power, Skip Buttons
I added power and skip diagrams, but only the skip diagram is used since the power button changes based on the symbol. I also added TTS for when those buttons are triggered.

I want to replace the buttons in the game with my custom ones next, and make the small audio and difficulty adjustments I've been meaning to make. I also want to add a whistle sound effect for the match start. After all that, there will be no more major changes.