Added Match Start SFX
I added a match start sound effect to the game. I don't think I'll add a "Ready..." sound effect to the game since I can't think of anything, and the timing is short anyway.

Now I'm going to make the AI easier and that should be all the major changes.