Replaced Arrows, Adjusted Difficulties, Adjusted Powers
I replaced the arrow buttons with just the arrows, since it looks better. All the arrow icon uses in the game should be updated and fixed so that they face the right directions.

I adjusted the difficulty of the A.I., including tweaking when they use their powers.

I adjusted how fast the powers fill up slightly. I also fixed a bug where the total power count had not been updated (this probably caused save issues, but if it did, I didn't notice).