Power None Desc, Computer AI, Power Twist, Equation Text Resizing
I rewrote the power none description to be a bit clearer. I also adjusted the twist power so that it runs out slightly faster.

The computer AI has been altered so that the gap to trigger a power is wider, making the AI easier.

I made the equation text box slightly bigger. I don't think that I can make the text any bigger, so hopefully it's okay.